move events2 code to events route (delele /events2)
