+++ frontend/src/app/(private)/events/page.tsx
@@ -1,11 +1,13 @@
 'use client';
 
-import { useEffect, useState, useCallback, Suspense } from 'react';
+import { Suspense } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { TabNav } from '@components/TabNav';
 import { EventCard } from '@components/EventCard';
 import { StickyButton } from '@components/StickyButton';
-import { eventsService, EventData } from '@services/eventsService';
+import { Button } from '@components/Button';
+import { useDashboardEvents, Relationship } from '@hooks/useDashboard';
+import { TabType as HookTabType } from '@hooks/useDashboard';
 import Image from 'next/image';
 import imageSrc from '../../../assets/images/empty-state.png';
 
@@ -17,119 +19,66 @@ const parseParamToTab = (param: string | null): TabType => {
   return 'ALL EVENTS';
 };
 
-const parseTabToParam = (tab: TabType): string => {
+const parseTabToParam = (tab: TabType): HookTabType => {
   if (tab === 'FUTURE EVENTS') return 'future';
   if (tab === 'ARCHIVED') return 'archived';
   return 'all';
 };
 
+type EventAction = 'archived' | 'edit' | 'leave' | 'join';
+const now = new Date();
+
+function getEventAction(
+  relationship: Relationship,
+  eventDate: Date,
+  currentTab: TabType
+): EventAction {
+  if (currentTab === 'ARCHIVED') return 'archived';
+
+  const isPast = eventDate < now;
+  if (isPast) return 'archived';
+
+  if (relationship === 'author') return 'edit';
+  if (relationship === 'joined') return 'leave';
+  return 'join';
+}
+
 function EventsDashboardContent() {
   const router = useRouter();
   const searchParams = useSearchParams();
 
-  // Tab synchronizing
   const currentTab = parseParamToTab(searchParams.get('tab'));
   const currentParam = parseTabToParam(currentTab);
 
-  const [events, setEvents] = useState<EventData[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-  const [pendingEventIds, setPendingEventIds] = useState<Set<string>>(new Set());
+  const { events, loading, join, leave, pending } = useDashboardEvents(currentParam);
 
-  const fetchDashboardEvents = useCallback(
-    async (isMounted: boolean = true) => {
-      try {
-        const data = await eventsService.getDashboardEvents(
-          currentParam as 'all' | 'future' | 'archived'
-        );
-        if (isMounted) {
-          const realEvents = Array.isArray(data) ? data : data?.events || [];
-          setEvents(realEvents);
-        }
-      } catch (err) {
-        console.error('Dashboard failed to retrieve view data:', err);
-      } finally {
-        if (isMounted) setIsLoading(false);
-      }
-    },
-    [currentParam]
-  );
-  // Update URL parameter
+  console.log('EVENTS: ', events);
+
+  // Tab change (only URL logic here)
   const handleTabChange = (selectedTab: string) => {
-    setIsLoading(true);
-    const paramValue = parseTabToParam(selectedTab as TabType);
     const params = new URLSearchParams(searchParams.toString());
-    params.set('tab', paramValue);
+    params.set('tab', parseTabToParam(selectedTab as TabType));
     router.push(`?${params.toString()}`);
   };
 
-  // Update event relationship (Join/Leave/Edit)
-  const handleEventAction = async (
-    eventId: string,
-    currentRelationship: 'author' | 'joined' | 'none'
-  ): Promise<void> => {
-    if (currentRelationship === 'author') {
-      router.push(`/events/${eventId}/edit`);
+  // Event action (thin controller)
+  const handleEventAction = (id: string, rel: Relationship) => {
+    if (rel === 'author') {
+      router.push(`/events/${id}/edit`);
       return;
     }
 
-    const isJoining = currentRelationship === 'none';
-
-    setPendingEventIds((prev) => new Set(prev).add(eventId));
-
-    try {
-      if (isJoining) {
-        await eventsService.joinEvent(eventId);
-      } else {
-        await eventsService.leaveEvent(eventId);
-      }
-      setEvents((prevEvents) =>
-        prevEvents.map((event) => {
-          if (event.id === eventId) {
-            return {
-              ...event,
-              relationship: isJoining ? 'joined' : 'none',
-              attendeeCount: isJoining
-                ? (event.attendeeCount ?? 0) + 1
-                : Math.max(0, (event.attendeeCount ?? 0) - 1),
-            };
-          }
-          return event;
-        })
-      );
-    } catch (error) {
-      console.error('Action failed:', error);
-      if (
-        error instanceof Error &&
-        (error.message === 'UNAUTHORIZED' || error.message.includes('token'))
-      ) {
-        router.push('/sign-in');
-        return;
-      }
-    } finally {
-      setPendingEventIds((prev) => {
-        const next = new Set(prev);
-        next.delete(eventId);
-        return next;
-      });
-    }
+    if (rel === 'none') join(id);
+    else leave(id);
   };
 
-  // Data fetching with loading state
-  useEffect(() => {
-    let isMounted = true;
-    const initializeDashboard = async () => {
-      await fetchDashboardEvents(isMounted);
-    };
-
-    initializeDashboard();
-
-    return () => {
-      isMounted = false;
-    };
-  }, [fetchDashboardEvents]);
+  const handleCreateEvent = () => {
+    router.push('/create-event');
+  };
 
   return (
-    <div className="w-full min-h-screen bg-gray-50 p-4 md:p-8 pb-24 md:pb-8 relative">
+    <div className="size-full flex flex-col bg-gray-50 p-4 md:p-8 pb-24 md:pb-8 relative">
+      {/* HEADER */}
       <header className="flex flex-col md:flex-row md:items-center md:justify-between gap-4 border-b border-gray-200 pb-4 mb-8 w-full">
         <div className="w-full md:w-auto">
           <TabNav
@@ -139,17 +88,16 @@ function EventsDashboardContent() {
           />
         </div>
 
-        <div className="w-full md:w-auto fixed bottom-0 left-0 right-0 p-4 bg-white border-t border-gray-200 md:static md:border-none md:p-0 md:bg-transparent z-50">
-          <StickyButton
-            label="CREATE NEW EVENT"
-            onClick={() => router.push('/events/create')}
-            className="w-full md:w-max block"
-          />
+        <div className="hidden sm:block">
+          <Button onClick={handleCreateEvent} size="default" variant="idle">
+            CREATE NEW EVENT
+          </Button>
         </div>
+        <StickyButton onClick={handleCreateEvent} label="CREATE NEW EVENT"></StickyButton>
       </header>
 
-      {/* 1. (LOADING STATE) */}
-      {isLoading && (
+      {/* LOADING */}
+      {loading && (
         <section className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 w-full">
           {Array.from({ length: 3 }).map((_, idx) => (
             <EventCard
@@ -162,9 +110,9 @@ function EventsDashboardContent() {
         </section>
       )}
 
-      {/* 2. (EMPTY STATE) */}
-      {!isLoading && events.length === 0 && (
-        <section className="flex flex-col items-center justify-center min-h-[300px] text-center gap-4 w-full">
+      {/* EMPTY */}
+      {!loading && events.length === 0 && (
+        <section className="flex-1 flex flex-col items-center justify-center text-center gap-4 w-full">
           <div>
             <Image
               src={imageSrc}
@@ -173,19 +121,20 @@ function EventsDashboardContent() {
               priority
             />
           </div>
+
           <h3 className="text-sm font-medium text-tabs-idle tracking-wide uppercase">
             THERE ARE NO EVENTS TO DISPLAY
           </h3>
         </section>
       )}
 
-      {/* 3. (LOADED STATE)*/}
-      {!isLoading && events.length > 0 && (
-        <section className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 w-full">
+      {/* LIST */}
+      {!loading && events.length > 0 && (
+        <section className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 w-full pb-20 sm:pb-8">
           {events.map((event) => (
             <EventCard
-              id={event.id}
               key={event.id}
+              id={event.id}
               isLoading={false}
               title={event.title}
               author={
@@ -195,16 +144,9 @@ function EventsDashboardContent() {
               description={event.description}
               attendeeCount={event.attendeeCount ?? 0}
               maxAttendees={event.maxAttendees ?? 50}
-              action={
-                currentTab === 'ARCHIVED'
-                  ? 'archived'
-                  : event.relationship === 'author'
-                    ? 'edit'
-                    : event.relationship === 'joined'
-                      ? 'leave'
-                      : 'join'
-              }
-              isActionPending={pendingEventIds.has(event.id)}
+              interactive={true}
+              action={getEventAction(event.relationship, new Date(event.date), currentTab)}
+              isActionPending={pending.has(event.id)}
               onActionClick={() => {
                 if (currentTab !== 'ARCHIVED') {
                   handleEventAction(event.id, event.relationship);
@@ -217,6 +159,7 @@ function EventsDashboardContent() {
     </div>
   );
 }
+
 export default function DashboardPage() {
   return (
     <Suspense
